Network postMapper

+++ app/src/main/java/ru/potemkin/orpheusjetpackcompose/data/model/CommentDto.kt
@@ -3,8 +3,8 @@ package ru.potemkin.orpheusjetpackcompose.data.model
 import com.google.gson.annotations.SerializedName
 
 data class CommentDto(
-    @SerializedName("id") val id: Long,
-    @SerializedName("from_id") val authorId: Long,
+    @SerializedName("id") val id: Int,
+    @SerializedName("from_id") val authorId: Int,
     @SerializedName("text") val text: String,
     @SerializedName("date") val date: Long
 )
